+++ einconv/kfac_reduce.py
@@ -153,7 +153,8 @@ def _kfac_reduce_factor_einsum_operands(
             padding=t_padding[n],
             dilation=t_dilation[n],
             device=input.device,
-        ).to(input.dtype)
+            dtype=input.dtype,
+        )
         for n in range(N)
     ]
